+++ src/command_modules/azure-cli-acr/azure/cli/command_modules/acr/custom.py
@@ -9,6 +9,7 @@ from knack.log import get_logger
 from azure.cli.core.commands import LongRunningOperation
 
 from azure.mgmt.containerregistry.v2017_10_01.models import (
+    Registry,
     RegistryUpdateParameters,
     StorageAccountProperties,
     SkuName,
@@ -19,7 +20,6 @@ from ._constants import MANAGED_REGISTRY_SKU
 from ._utils import (
     arm_deploy_template_new_storage,
     arm_deploy_template_existing_storage,
-    arm_deploy_template_managed_storage,
     random_storage_account_name,
     get_registry_by_name,
     validate_managed_registry,
@@ -91,23 +91,14 @@ def acr_create(cmd,
                     admin_user_enabled,
                     deployment_name)
             )
+        return client.get(resource_group_name, registry_name)
     else:
         if storage_account_name:
             logger.warning(
                 "The registry '%s' in '%s' SKU is a managed registry. The specified storage account will be ignored.",
                 registry_name, sku)
-        LongRunningOperation(cmd.cli_ctx)(
-            arm_deploy_template_managed_storage(
-                cmd.cli_ctx,
-                resource_group_name,
-                registry_name,
-                location,
-                sku,
-                admin_user_enabled,
-                deployment_name)
-        )
-
-    return client.get(resource_group_name, registry_name)
+        registry = Registry(location=location, sku=Sku(name=sku), admin_user_enabled=admin_user_enabled)
+        return client.create(resource_group_name, registry_name, registry)
 
 
 def acr_delete(cmd, client, registry_name, resource_group_name=None):
@@ -169,20 +160,8 @@ def acr_login(cmd, registry_name, resource_group_name=None, username=None, passw
     if in_cloud_console():
         raise CLIError('This command requires running the docker daemon, which is not supported in Azure Cloud Shell.')
 
-    from subprocess import PIPE, Popen, CalledProcessError
-    docker_not_installed = "Please verify if docker is installed."
-    docker_not_available = "Please verify if docker daemon is running properly."
-
-    try:
-        p = Popen(["docker", "ps"], stdout=PIPE, stderr=PIPE)
-        _, stderr = p.communicate()
-    except OSError:
-        raise CLIError(docker_not_installed)
-    except CalledProcessError:
-        raise CLIError(docker_not_available)
-
-    if stderr:
-        raise CLIError(stderr.decode())
+    from subprocess import PIPE, Popen
+    docker_command = _get_docker_command()
 
     login_server, username, password = get_login_credentials(
         cli_ctx=cmd.cli_ctx,
@@ -191,46 +170,23 @@ def acr_login(cmd, registry_name, resource_group_name=None, username=None, passw
         username=username,
         password=password)
 
-    use_password_stdin = False
-    try:
-        p = Popen(["docker", "login", "--help"], stdout=PIPE, stderr=PIPE)
-        stdout, _ = p.communicate()
-        if b'--password-stdin' in stdout:
-            use_password_stdin = True
-    except OSError:
-        raise CLIError(docker_not_installed)
-    except CalledProcessError:
-        raise CLIError(docker_not_available)
-
-    if use_password_stdin:
-        p = Popen(["docker", "login",
-                   "--username", username,
-                   "--password-stdin",
-                   login_server], stdin=PIPE, stderr=PIPE)
-        _, stderr = p.communicate(input=password.encode())
-    else:
-        p = Popen(["docker", "login",
-                   "--username", username,
-                   "--password", password,
-                   login_server], stderr=PIPE)
-        _, stderr = p.communicate()
+    p = Popen([docker_command, "login",
+               "--username", username,
+               "--password", password,
+               login_server], stderr=PIPE)
+    _, stderr = p.communicate()
 
     if stderr:
         if b'error storing credentials' in stderr and b'stub received bad data' in stderr \
            and _check_wincred(login_server):
             # Retry once after disabling wincred
-            if use_password_stdin:
-                p = Popen(["docker", "login",
-                           "--username", username,
-                           "--password-stdin",
-                           login_server], stdin=PIPE)
-                p.communicate(input=password.encode())
-            else:
-                p = Popen(["docker", "login",
-                           "--username", username,
-                           "--password", password,
-                           login_server])
-                p.wait()
+            p = Popen([docker_command, "login",
+                       "--username", username,
+                       "--password", password,
+                       login_server])
+            p.wait()
+        elif b'--password-stdin' in stderr:
+            pass
         else:
             import sys
             output = getattr(sys.stderr, 'buffer', sys.stderr)
@@ -243,6 +199,34 @@ def acr_show_usage(cmd, client, registry_name, resource_group_name=None):
                                                        resource_group_name,
                                                        "Usage is only supported for managed registries.")
     return client.list_usages(resource_group_name, registry_name)
+
+
+def _get_docker_command():
+    docker_not_installed = "Please verify if docker is installed."
+    docker_not_available = "Please verify if docker daemon is running properly."
+    docker_command = 'docker'
+
+    from subprocess import PIPE, Popen, CalledProcessError
+    try:
+        p = Popen([docker_command, "ps"], stdout=PIPE, stderr=PIPE)
+        _, stderr = p.communicate()
+    except OSError:
+        # docker is not discoverable in WSL so retry docker.exe once
+        try:
+            docker_command = 'docker.exe'
+            p = Popen([docker_command, "ps"], stdout=PIPE, stderr=PIPE)
+            _, stderr = p.communicate()
+        except OSError:
+            raise CLIError(docker_not_installed)
+        except CalledProcessError:
+            raise CLIError(docker_not_available)
+    except CalledProcessError:
+        raise CLIError(docker_not_available)
+
+    if stderr:
+        raise CLIError(stderr.decode())
+
+    return docker_command
 
 
 def _check_wincred(login_server):