+++ src/command_modules/azure-cli-vm/azure/cli/command_modules/vm/tests/latest/test_vm_commands.py
@@ -1951,43 +1951,23 @@ class MSIScenarioTest(ScenarioTest):
             self.cmd('vm create -g {rg} -n {vm1} --image debian --assign-identity --admin-username admin123 --admin-password PasswordPassword1! --scope {scope}', checks=[
                 self.check('identity.role', 'Contributor'),
                 self.check('identity.scope', '/subscriptions/{sub}/resourceGroups/{rg}'),
-                self.check('identity.port', 50342)
-            ])
-
-            self.cmd('vm extension list -g {rg} --vm-name {vm1}', checks=[
-                self.check('[0].virtualMachineExtensionType', 'ManagedIdentityExtensionForLinux'),
-                self.check('[0].settings.port', 50342)
             ])
 
             # create a windows vm with reader role on the linux vm
             result = self.cmd('vm create -g {rg} -n {vm2} --image Win2016Datacenter --assign-identity --scope {vm1_id} --role reader --admin-username admin123 --admin-password PasswordPassword1!', checks=[
                 self.check('identity.role', 'reader'),
                 self.check('identity.scope', '{vm1_id}'),
-                self.check('identity.port', 50342)
             ])
             uuid.UUID(result.get_output_in_json()['identity']['systemAssignedIdentity'])
-
-            self.cmd('vm extension list -g {rg} --vm-name {vm2}', checks=[
-                self.check('[0].virtualMachineExtensionType', 'ManagedIdentityExtensionForWindows'),
-                self.check('[0].publisher', 'Microsoft.ManagedIdentity'),
-                self.check('[0].settings.port', 50342)
-            ])
 
             # create a linux vm w/o identity and later enable it
             vm3_result = self.cmd('vm create -g {rg} -n {vm3} --image debian --admin-username admin123 --admin-password PasswordPassword1!').get_output_in_json()
             self.assertIsNone(vm3_result.get('identity'))
-            result = self.cmd('vm identity assign -g {rg} -n {vm3} --scope {vm1_id} --role reader --port 50343', checks=[
+            result = self.cmd('vm identity assign -g {rg} -n {vm3} --scope {vm1_id} --role reader', checks=[
                 self.check('role', 'reader'),
                 self.check('scope', '{vm1_id}'),
-                self.check('port', 50343)
             ])
             uuid.UUID(result.get_output_in_json()['systemAssignedIdentity'])
-
-            self.cmd('vm extension list -g {rg} --vm-name {vm3}', checks=[
-                self.check('[0].virtualMachineExtensionType', 'ManagedIdentityExtensionForLinux'),
-                self.check('[0].publisher', 'Microsoft.ManagedIdentity'),
-                self.check('[0].settings.port', 50343)
-            ])
 
     @ResourceGroupPreparer(name_prefix='cli_test_vmss_msi')
     def test_vmss_msi(self, resource_group):
@@ -2006,44 +1986,24 @@ class MSIScenarioTest(ScenarioTest):
             self.cmd('vmss create -g {rg} -n {vmss1} --image debian --instance-count 1 --assign-identity --admin-username admin123 --admin-password PasswordPassword1! --scope {scope}', checks=[
                 self.check('vmss.identity.role', 'Contributor'),
                 self.check('vmss.identity.scope', '/subscriptions/{sub}/resourceGroups/{rg}'),
-                self.check('vmss.identity.port', 50342)
-            ])
-
-            self.cmd('vmss extension list -g {rg} --vmss-name {vmss1}', checks=[
-                self.check('[0].type', 'ManagedIdentityExtensionForLinux'),
-                self.check('[0].publisher', 'Microsoft.ManagedIdentity'),
-                self.check('[0].settings.port', 50342)
             ])
 
             # create a windows vm with reader role on the linux vm
             result = self.cmd('vmss create -g {rg} -n {vmss2} --image Win2016Datacenter --instance-count 1 --assign-identity --scope {vmss1_id} --role reader --admin-username admin123 --admin-password PasswordPassword1!', checks=[
                 self.check('vmss.identity.role', 'reader'),
                 self.check('vmss.identity.scope', '{vmss1_id}'),
-                self.check('vmss.identity.port', 50342)
             ]).get_output_in_json()
             uuid.UUID(result['vmss']['identity']['systemAssignedIdentity'])
-
-            self.cmd('vmss extension list -g {rg} --vmss-name {vmss2}', checks=[
-                self.check('[0].type', 'ManagedIdentityExtensionForWindows'),
-                self.check('[0].publisher', 'Microsoft.ManagedIdentity'),
-                self.check('[0].settings.port', 50342)
-            ])
 
             # create a linux vm w/o identity and later enable it
             result = self.cmd('vmss create -g {rg} -n {vmss3} --image debian --instance-count 1 --admin-username admin123 --admin-password PasswordPassword1!').get_output_in_json()['vmss']
             self.assertIsNone(result.get('identity'))
 
-            result = self.cmd('vmss identity assign -g {rg} -n {vmss3} --scope "{vmss1_id}" --role reader --port 50343', checks=[
+            result = self.cmd('vmss identity assign -g {rg} -n {vmss3} --scope "{vmss1_id}" --role reader', checks=[
                 self.check('role', 'reader'),
                 self.check('scope', '{vmss1_id}'),
-                self.check('port', 50343)
             ]).get_output_in_json()
             uuid.UUID(result['systemAssignedIdentity'])
-
-            self.cmd('vmss extension list -g {rg} --vmss-name {vmss3}', checks=[
-                self.check('[0].type', 'ManagedIdentityExtensionForLinux'),
-                self.check('[0].settings.port', 50343)
-            ])
 
     @ResourceGroupPreparer(name_prefix='cli_test_msi_no_scope')
     def test_msi_no_scope(self, resource_group):
@@ -2059,24 +2019,11 @@ class MSIScenarioTest(ScenarioTest):
         self.cmd('vm create -g {rg} -n {vm1} --image debian --assign-identity --admin-username admin123 --admin-password PasswordPassword1!', checks=[
             self.check('identity.scope', None),
             self.check('identity.role', None),
-            self.check('identity.port', 50342)
-        ])
-        # the extension should still get provisioned
-        self.cmd('vm extension list -g {rg} --vm-name {vm1}', checks=[
-            self.check('[0].virtualMachineExtensionType', 'ManagedIdentityExtensionForLinux'),
-            self.check('[0].settings.port', 50342)
         ])
 
         # create a vmss with identity but w/o a role assignment (--scope "")
         self.cmd('vmss create -g {rg} -n {vmss1} --image debian --assign-identity --admin-username admin123 --admin-password PasswordPassword1!', checks=[
             self.check('vmss.identity.scope', None),
-            self.check('vmss.identity.port', 50342)
-        ])
-
-        # the extension should still get provisioned
-        self.cmd('vmss extension list -g {rg} --vmss-name {vmss1}', checks=[
-            self.check('[0].type', 'ManagedIdentityExtensionForLinux'),
-            self.check('[0].settings.port', 50342)
         ])
 
         # create a vm w/o identity
@@ -2084,23 +2031,11 @@ class MSIScenarioTest(ScenarioTest):
         # assign identity but w/o a role assignment
         self.cmd('vm identity assign -g {rg} -n {vm2}', checks=[
             self.check('scope', None),
-            self.check('port', 50342)
-        ])
-        # the extension should still get provisioned
-        self.cmd('vm extension list -g {rg} --vm-name {vm2}', checks=[
-            self.check('[0].virtualMachineExtensionType', 'ManagedIdentityExtensionForLinux'),
-            self.check('[0].settings.port', 50342)
         ])
 
         self.cmd('vmss create -g {rg} -n {vmss2} --image debian --admin-username admin123 --admin-password PasswordPassword1!')
         self.cmd('vmss identity assign -g {rg} -n {vmss2}', checks=[
             self.check('scope', None),
-            self.check('port', 50342)
-        ])
-
-        self.cmd('vmss extension list -g {rg} --vmss-name {vmss2}', checks=[
-            self.check('[0].type', 'ManagedIdentityExtensionForLinux'),
-            self.check('[0].settings.port', 50342)
         ])
 
     @ResourceGroupPreparer(random_name_length=20, location='westcentralus')
@@ -2393,17 +2328,17 @@ class VMRunCommandScenarioTest(ScenarioTest):
         self.cmd('vm run-command invoke -g {rg} -n{vm} --command-id RunShellScript  --scripts "echo $0 $1" --parameters hello world')
 
 
-# TOOD: FAIL no KeyVault commands!
 @api_version_constraint(ResourceType.MGMT_COMPUTE, min_api='2017-03-30')
-class VMSSDiskEncryptionTest(ScenarioTest):
+class VMDiskEncryptionTest(ScenarioTest):
 
-    @ResourceGroupPreparer(name_prefix='cli_test_vmss_encryption', location='eastus2euap')  # the feature is only available in canary, should rollout to public soon
+    @ResourceGroupPreparer(name_prefix='cli_test_vmss_encryption', location='westus')
     def test_vmss_disk_encryption_e2e(self, resource_group, resource_group_location):
         self.kwargs.update({
             'vault': self.create_random_name('vault', 10),
             'vmss': 'vmss1'
         })
         self.cmd('keyvault create -g {rg} -n {vault} --enabled-for-disk-encryption "true"')
+        time.sleep(60)  # to avoid 504(too many requests) on a newly created vault
         self.cmd('vmss create -g {rg} -n {vmss} --image win2016datacenter --instance-count 1 --admin-username clitester1 --admin-password Test123456789!')
         self.cmd('vmss encryption enable -g {rg} -n {vmss} --disk-encryption-keyvault {vault}')
         self.cmd('vmss update-instances -g {rg} -n {vmss}  --instance-ids "*"')
@@ -2419,6 +2354,19 @@ class VMSSDiskEncryptionTest(ScenarioTest):
             self.check('virtualMachineProfile.extensionProfile.extensions[0].settings.EncryptionOperation', 'DisableEncryption'),
             self.check('virtualMachineProfile.extensionProfile.extensions[0].settings.VolumeType', 'ALL')
         ])
+
+    @ResourceGroupPreparer(name_prefix='cli_test_vm_encryption', location='westus')
+    def test_vm_disk_encryption_e2e(self, resource_group, resource_group_location):
+        self.kwargs.update({
+            'vault': self.create_random_name('vault', 10),
+            'vm': 'vm1'
+        })
+        self.cmd('keyvault create -g {rg} -n {vault} --enabled-for-disk-encryption "true"')
+        time.sleep(60)  # to avoid 504(too many requests) on a newly created vault
+        self.cmd('vm create -g {rg} -n {vm} --image win2012datacenter --admin-username clitester1 --admin-password Test123456789!')
+        self.cmd('vm encryption enable -g {rg} -n {vm} --disk-encryption-keyvault {vault}')
+        self.cmd('vm encryption show -g {rg} -n {vm}', checks=[self.check('disks[0].statuses[0].code', 'EncryptionState/encrypted')])
+        self.cmd('vm encryption disable -g {rg} -n {vm}')
 
 
 @api_version_constraint(ResourceType.MGMT_COMPUTE, min_api='2017-03-30')