+++ src/command_modules/azure-cli-vm/HISTORY.rst
@@ -4,6 +4,9 @@ Release History
 ===============
 2.0.31
 ++++++
+* vm: fix an invalid detection logic on unmanaged blob uri
+* vm: support disk encryption w/o user provided service principals 
+* BREAKING CHANGE: do not use VM 'ManagedIdentityExtension' for MSI support
 * vmss: support eviction policy
 * BREAKING CHANGE: remove erroneous argument of `ids` from `vm extension list`,
                    `vm secret list`, `vm unmanaged-disk list` and  `vmss nic list`